+++ config/silverstripe-5-1.php
@@ -2,9 +2,17 @@
 
 declare(strict_types=1);
 
+use Netwerkstatt\SilverstripeRector\Rector\Misc\PropertyToConfigRector;
 use Rector\Config\RectorConfig;
 
 return static function (RectorConfig $rectorConfig): void {
-
-
+    $rectorConfig->ruleWithConfiguration(PropertyToConfigRector::class, [
+        'SilverStripe\Dev\BuildTask' => [
+            'enabled' => [
+                'name' => 'is_enabled',
+                'visibility' => \PhpParser\Modifiers::PRIVATE | \PhpParser\Modifiers::STATIC,
+                'add_config' => true,
+            ],
+        ],
+    ]);
 };